+++ widgets/DownloadIcsLink.php
@@ -11,7 +11,6 @@ use humhub\components\Widget;
 use humhub\libs\Html;
 use humhub\modules\calendar\helpers\Url;
 use humhub\modules\calendar\interfaces\event\CalendarEventIF;
-use humhub\modules\calendar\models\CalendarEntry;
 use Yii;
 
 /**
@@ -28,9 +27,17 @@ class DownloadIcsLink extends Widget
     public function run()
     {
         if ($this->calendarEntry === null) {
-            return;
+            return '';
         }
 
-        return Html::a(Yii::t('CalendarModule.base', 'Download ICS'), Url::toEntryDownloadICS($this->calendarEntry), ['target' => '_blank']);
+        return Html::tag(
+            'span',
+            Html::a(
+                Yii::t('CalendarModule.base', 'Download ICS'),
+                Url::toEntryDownloadICS($this->calendarEntry),
+                ['target' => '_blank'],
+            ),
+            ['class' => 'calendar-entry-ics-download'],
+        );
     }
 }